Parallelization of the code with OpenMP
The exploration loop is now trivially parallelized using OpenMP.
This was not as easy as expected since lap.f90 had to be modified to work in parallel.
Note: Note that variables which are explicitly initialized in declarations or in DATA statements automatically acquire the SAVE attribute.
This means the variable continues where it was the next time it enters the precedure and the same address is used in memory for the variable in all instances of the procedure.