Multiple Atom Unit Cell Implementation + Slanting problem
The code can now run and display multiple atom unit cells starting from pylada structures. Most of the work as been done on the displaying routines, the optimization already worked with multiple atoms as it was.

I added an extra step at the end of fastoptimization where the distance suqared is optimized, that solves the slanting problem as slanting is no longer advantageous.