Bugs are fixed in 3D implementation - SLANTING PROBLEM
Fixed most bugs in 3D implementation, the rotation is now controlled by 3 angles instead of 1 angles and 2 vector components. It caused discontinuities in the grandient denscent which made it fail.

Most of the display is updated to work in 3D, gif animations are created to help 3D visualisation.

The slanting problem is still unsolved and causes the classification routine to fail. The current version of the code has a raise error due to on going testing of the slanting.