Working version of interface matching
The code give a unit cell with both atoms types (substrate and material) in addition to the displacement cell.
prepare_poscar.py creates poscars from the poscars of the surface cutting algorithm using the cell and the transformations from the mapping algorithm.

The code should start directly from the surface cutting poscars and directly output the new oriented poscars.