Added animatioin making in p2ptrans, remaing in calss. and explore free in transform
p2ptrans creates animations of the transition. It also creates poscars of the final structure and the stretched initial structure. It tries to give stretching direction but it's not trusthworthy. The creation of poscars at every step of the transition is being implemented but untested.

Transform is modified to be able to remap during the classification using the remap flag in the namelist. The calssification criteria is alos changed slighlty. The free explore routine is added (for Diamond to graphite for example).