Several bug fixes from local changes
Fixed a bug with OR in analysis, was using transpose to find the inverse but this is true only if the vectors are normalized first.
Added hability to specify HP produceTrans in core.py and config.py
Slab with c not along z are now allowed in interfaces.py
Make sure the interface is not flipped in _gd_vec in transform.f90
Changed the minor version. Because of changes in previous commits dmin from previous runs will create errors.